+++ tests/test_visualization_molecules.py
@@ -1,6 +1,8 @@
 import os
 
 import numpy as np
+import matplotlib.pyplot as plt
+from imageio.v2 import imread
 
 from visual_graph_datasets.visualization.base import create_frameless_figure
 from visual_graph_datasets.visualization.molecules import mol_from_smiles
@@ -28,14 +30,23 @@ def test_visualize_molecular_graph_basically_works():
     assert isinstance(svg_string, str)
     assert len(svg_string) != 0
 
+    # Here we are going to save it as a PNG file and then afterwards load it again
+    # using imread to emulate the process of visualization as best as possible
+    vis_path = os.path.join(ARTIFACTS_PATH, 'molecule_visualization.png')
+    fig.savefig(vis_path)
+    plt.close(fig)
+
+    fig, ax = plt.subplots(ncols=1, nrows=1, figsize=(10, 10))
+    image = imread(vis_path)
+    ax.imshow(image)
+
     # to make sure the positions are correct we are going to scatter a small dot to every position and can
     # then check in the image if they are at the correct atom positions.
     for i in range(num_atoms):
         ax.scatter(
             *node_positions[i],
-            color='white',
+            color='red',
         )
 
-    vis_path = os.path.join(ARTIFACTS_PATH, 'molecule_visualization.png')
-    fig.savefig(vis_path)
-
+    img_path = os.path.join(ARTIFACTS_PATH, 'molecule_visualization_loaded.png')
+    fig.savefig(img_path)